+++ src/gwt/src/org/rstudio/core/client/widget/GridViewerStyles.java
@@ -22,45 +22,45 @@ public class GridViewerStyles
    public static String getCustomStyle()
    {
       return
-         ".rstudio-themes-flat.rstudio-themes-default #rsGridData_info,\n" +
-         ".rstudio-themes-flat.rstudio-themes-default table.dataTable thead th,\n" +
-         ".rstudio-themes-flat.rstudio-themes-default table.dataTable thead td,\n" +
-         ".rstudio-themes-flat.rstudio-themes-default table.dataTable tbody td {\n" +
+         ".rstudio-themes-default #rsGridData_info,\n" +
+         ".rstudio-themes-default table.dataTable thead th,\n" +
+         ".rstudio-themes-default table.dataTable thead td,\n" +
+         ".rstudio-themes-default table.dataTable tbody td {\n" +
          "   border-color: " + ThemeColors.defaultBorder + ";\n" +
          "}\n" +
          "\n" +
-         ".rstudio-themes-flat.rstudio-themes-dark-grey #rsGridData_info,\n" +
-         ".rstudio-themes-flat.rstudio-themes-dark-grey table.dataTable thead th,\n" +
-         ".rstudio-themes-flat.rstudio-themes-dark-grey table.dataTable thead td,\n" +
-         ".rstudio-themes-flat.rstudio-themes-dark-grey table.dataTable tbody td {\n" +
+         ".rstudio-themes-dark-grey #rsGridData_info,\n" +
+         ".rstudio-themes-dark-grey table.dataTable thead th,\n" +
+         ".rstudio-themes-dark-grey table.dataTable thead td,\n" +
+         ".rstudio-themes-dark-grey table.dataTable tbody td {\n" +
          "   border-color: " + ThemeColors.darkGreyBorder + ";\n" +
          "}\n" +
          "\n" +
-         ".rstudio-themes-flat.rstudio-themes-alternate #rsGridData_info,\n" +
-         ".rstudio-themes-flat.rstudio-themes-alternate table.dataTable thead th,\n" +
-         ".rstudio-themes-flat.rstudio-themes-alternate table.dataTable thead td,\n" +
-         ".rstudio-themes-flat.rstudio-themes-alternate table.dataTable tbody td {\n" +
+         ".rstudio-themes-alternate #rsGridData_info,\n" +
+         ".rstudio-themes-alternate table.dataTable thead th,\n" +
+         ".rstudio-themes-alternate table.dataTable thead td,\n" +
+         ".rstudio-themes-alternate table.dataTable tbody td {\n" +
          "   border-color: " + ThemeColors.alternateBorder + ";\n" +
          "}\n" +
          "\n" +
-         ".rstudio-themes-flat.rstudio-themes-default #rsGridData_info,\n" +
-         ".rstudio-themes-flat.rstudio-themes-default th,\n" +
-         ".rstudio-themes-flat.rstudio-themes-default td.first-child,\n" +
-         ".rstudio-themes-flat.rstudio-themes-default table.dataTable thead td {\n" +
+         ".rstudio-themes-default #rsGridData_info,\n" +
+         ".rstudio-themes-default th,\n" +
+         ".rstudio-themes-default td.first-child,\n" +
+         ".rstudio-themes-default table.dataTable thead td {\n" +
          "   background-color: " + ThemeColors.defaultBackground + ";\n" +
          "}\n" +
          "\n" +
-         ".rstudio-themes-flat.rstudio-themes-dark-grey #rsGridData_info,\n" +
-         ".rstudio-themes-flat.rstudio-themes-dark-grey th,\n" +
-         ".rstudio-themes-flat.rstudio-themes-dark-grey td.first-child,\n" +
-         ".rstudio-themes-flat.rstudio-themes-dark-grey table.dataTable thead th {\n" +
+         ".rstudio-themes-dark-grey #rsGridData_info,\n" +
+         ".rstudio-themes-dark-grey th,\n" +
+         ".rstudio-themes-dark-grey td.first-child,\n" +
+         ".rstudio-themes-dark-grey table.dataTable thead th {\n" +
          "   background-color: " + ThemeColors.darkGreyBackground + ";\n" +
          "}\n" +
          "\n" +
-         ".rstudio-themes-flat.rstudio-themes-alternate #rsGridData_info,\n" +
-         ".rstudio-themes-flat.rstudio-themes-alternate th,\n" +
-         ".rstudio-themes-flat.rstudio-themes-alternate td.first-child,\n" +
-         ".rstudio-themes-flat.rstudio-themes-alternate table.dataTable thead th {\n" +
+         ".rstudio-themes-alternate #rsGridData_info,\n" +
+         ".rstudio-themes-alternate th,\n" +
+         ".rstudio-themes-alternate td.first-child,\n" +
+         ".rstudio-themes-alternate table.dataTable thead th {\n" +
          "   background-color: " + ThemeColors.alternateBackground + ";\n" +
          "}\n" +
          "\n";