+++ src/gwt/src/org/rstudio/studio/client/workbench/views/environment/EnvironmentPane.java
@@ -270,10 +270,6 @@ public class EnvironmentPane extends WorkbenchPane
          }
       });
 
-      if (!RStudioThemes.isFlat(prefs_)) {
-         searchWidget.getElement().getStyle().setMarginTop(1, Unit.PX);
-      }
-
       toolbar.addRightWidget(searchWidget);
 
       secondaryToolbar_ = toolbar;